refactor: code review changes
