+++ src/helpers/getTextContent.ts
