+++ src/helpers/matchers/matchTextContent.ts
@@ -2,7 +2,7 @@ import { Text } from 'react-native';
 import type { ReactTestInstance } from 'react-test-renderer';
 import { matches, TextMatch, TextMatchOptions } from '../../matches';
 import { filterNodeByType } from '../filterNodeByType';
-import { getTextContent } from '../textContent';
+import { getTextContent } from '../getTextContent';
 
 export function matchTextContent(
   node: ReactTestInstance,
